feat(dpi): rasterize glyphs at device resolution
Slice 1 laid out in logical pixels but still shaped and rasterized glyphs at
their logical size, so text on a HiDPI display was a logical-size bitmap
stretched to fill — correctly placed but soft.

Hand the scale factor to `LayoutGlyph::physical`, which snaps each glyph to the
physical pixel grid and stamps `font_size * scale` into its cache key, so the
rasterizer produces a bitmap at device resolution. Shaping itself stays logical:
metrics, wrap width, caret geometry and the reported extent are untouched, so a
DPI change can neither re-wrap a paragraph nor move a caret. Glyph positions come
back divided into logical space; the renderer divides the bitmap's physical
bearings and extent by the scale when it builds the quad, and the NDC map
multiplies the whole thing back — a round trip that lands each bitmap on the
exact device pixels it was cut for.

Two things this had to get right, each with a test that fails without it:

- The shape cache key now folds in the scale. Two displays produce identical
  layout but different bitmaps, so without this a cached entry hands back the
  wrong resolution — text stuck soft with nothing visibly wrong.
- The baseline offset handed to `physical()` is pre-scaled. That function scales
  the glyph's own position but adds the offset as-is (it wants a physical pen
  origin), so a logical baseline came back halved after the divide and dragged
  every line ~7px up its box. Invisible at x, where the offset is zero, which is
  why the x-only invariance test missed it.

Glyph coordinates crossing the render boundary are now logical f32 rather than
physical i32, which drops the `as i32` casts at every draw_glyph call site.

+++ crates/shroud_text/tests/text_engine_tests.rs
@@ -46,7 +46,7 @@ fn shape_text_glyph_y_includes_baseline() {
     // be at least 10 px below the block top; near-zero means the offset is lost.
     let baseline = result.glyphs[0].y;
     assert!(
-        baseline > 10,
+        baseline > 10.0,
         "glyph.y ({}) looks like we forgot to add line_y — should be the baseline Y",
         baseline
     );
@@ -770,7 +770,7 @@ fn shape_rich_underline_sits_below_the_baseline() {
         "one underline on one line, got {:?}",
         shaped.decoration_lines
     );
-    let baseline = shaped.glyphs[0].y as f32;
+    let baseline = shaped.glyphs[0].y;
     let line = shaped.decoration_lines[0];
     assert!(
         line.rect.origin.y > baseline,
@@ -796,7 +796,7 @@ fn shape_rich_strikethrough_crosses_above_the_baseline() {
     let mut engine = TextEngine::new();
     let shaped = engine.shape_rich(&[TextSpan::new("Hello").strikethrough()], 32.0, 40.0, None);
     assert_eq!(shaped.decoration_lines.len(), 1);
-    let baseline = shaped.glyphs[0].y as f32;
+    let baseline = shaped.glyphs[0].y;
     let line = shaped.decoration_lines[0];
     assert!(
         line.rect.origin.y < baseline,
