+++ crates/shroud_widgets/src/checkbox.rs
@@ -244,8 +244,8 @@ impl Widget for Checkbox {
                 for glyph in &shaped.glyphs {
                     if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                         ctx.draw_glyph(
-                            label_x as i32 + glyph.x,
-                            label_y as i32 + glyph.y,
+                            label_x + glyph.x,
+                            label_y + glyph.y,
                             image,
                             label_color,
                             glyph.cache_key,
