+++ crates/shroud_widgets/src/secure_input.rs
@@ -619,8 +619,8 @@ impl Widget for SecureInput {
                 for glyph in &shaped.glyphs {
                     if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                         ctx.draw_glyph(
-                            text_x as i32 + glyph.x,
-                            text_y as i32 + glyph.y,
+                            text_x + glyph.x,
+                            text_y + glyph.y,
                             image,
                             placeholder_color,
                             glyph.cache_key,
@@ -650,8 +650,8 @@ impl Widget for SecureInput {
                     for glyph in &shaped.glyphs {
                         if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                             ctx.draw_secure_glyph(
-                                text_x as i32 + glyph.x,
-                                text_y as i32 + glyph.y,
+                                text_x + glyph.x,
+                                text_y + glyph.y,
                                 image,
                                 text_color,
                                 glyph.cache_key,
@@ -711,8 +711,8 @@ impl Widget for SecureInput {
                 for glyph in &shaped.glyphs {
                     if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                         ctx.draw_glyph(
-                            text_x as i32 + glyph.x,
-                            text_y as i32 + glyph.y,
+                            text_x + glyph.x,
+                            text_y + glyph.y,
                             image,
                             text_color,
                             glyph.cache_key,
