+++ crates/shroud_widgets/src/segmented.rs
@@ -363,8 +363,8 @@ impl Widget for Segmented {
             for glyph in &shaped.glyphs {
                 if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                     ctx.draw_glyph(
-                        text_x as i32 + glyph.x,
-                        text_y as i32 + glyph.y,
+                        text_x + glyph.x,
+                        text_y + glyph.y,
                         image,
                         color,
                         glyph.cache_key,
